feat: add Containerfile for image builds
The image is based on coreos-installer for support of directly outputting ISO
images and includes the *butane* binary for rendering of configs.

The package is installed via pip as editable to automatically install
dependencies from the pyproject.toml and avoid the need to reinstall them when
updating pyromaniacs source code.

The /spec, /data/secrets, and /data/cache volumes will be used for mounting
pyromaniac configurations, HTTP(S) secrets when serving configs over the
network, and for caching ISO base images respectively.

Port 8000 is exposed for serving configs over HTTP(S).