fix: rewrite execution context to be pure dictionary
Pythons `exec(...)` function explicitly only supports pure dicts as
execution contexts. The context is now constructed in a simple one-liner
function.

The Library class is rewritten to be a proper mapping and support
unpacking for the context creation. The slash sugar for relative file
paths is moved to the View class and therefore doesn't only work for the
underscore variable anymore but also after arbitrary traversal of the
component tree.

The compiler and component modules are adapted to the new context
creation.