feat: add user friendly errors and error class documantation
Implements informative errors messages for all error classes.

A component call stack is added to the CompilerError class and code is
added to populate this stack when executing components.

The InvalidSignatureError from the *segment* subpackage is renamed to
SignatureSyntaxError to avoid collisions with the equally named error
from the *code* subpackage.

Docstrings are added to all error classes.