refactor: rework server address and auth handling
Renames the "Pyromaniac" class to "Remote" and moves its instantiation
into the __main__ module. Also moves the secret generation to when it is
actually first needed simplifying the argument parsing.

Adds embedding of the root certificate in the remote ISO generation and
switches to an ignition config merge instead of config replace to keep
the certificate active for the definitions from the remote file.