refactor: make compatible with Astral tooling
Astral's uv wants a minimal supported python version configured. The
ruff and ty linters/language servers can also be configured to honor
this version and help make sure the code stays compatible to it. This
commit configures python version 3.11 for all of them and adds some
uv-specific lines to the .gitignore file.